use master.password file and format on freebsd platform
